feat(types): standardize value handling with Rao newtype across SDK
Introduce u32 casts for voting fields (index, threshold, end) to match ProposalVoteData types. Standardize balance and value handling across queries, types, CLI commands, and tests using the Rao newtype. Update delegate, neuron, subnet, and stake types for consistent value representation.

+++ src/queries/voting.rs
@@ -24,15 +24,15 @@ pub async fn get_vote_data(
         let index = fields
             .get("index")
             .and_then(|v| decode_u64(v).ok())
-            .unwrap_or(0);
+            .unwrap_or(0) as u32;
         let threshold = fields
             .get("threshold")
             .and_then(|v| decode_u64(v).ok())
-            .unwrap_or(0);
+            .unwrap_or(0) as u32;
         let end = fields
             .get("end")
             .and_then(|v| decode_u64(v).ok())
-            .unwrap_or(0);
+            .unwrap_or(0) as u32;
 
         let ayes = fields
             .get("ayes")
