+++ src/types/proposal_vote.rs
@@ -3,11 +3,11 @@ use sp_core::crypto::AccountId32;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct ProposalVoteData {
-    pub index: u64,
-    pub threshold: u64,
+    pub index: u32,
+    pub threshold: u32,
     #[serde(with = "crate::utils::ss58::serde_account_vec")]
     pub ayes: Vec<AccountId32>,
     #[serde(with = "crate::utils::ss58::serde_account_vec")]
     pub nays: Vec<AccountId32>,
-    pub end: u64,
+    pub end: u32,
 }
